requirements.txt
APScheduler==3.11.2
beautifulsoup4==4.14.3
chromadb==1.5.2
fastapi==0.135.1
langchain==1.2.10
langchain-anthropic==1.3.4
langchain-chroma==1.1.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
openai==2.26.0
pydantic-settings==2.13.1
python-dotenv==1.2.2
requests==2.32.5
tiktoken==0.12.0
uvicorn[standard]==0.41.0

